Added server plugin (Linux & Windows)
Fixed major bugs
Improved network by implenting boost::asio
Added .asi blocking & some security checks

+++ client-loader/loader.cpp
@@ -18,12 +18,20 @@ BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved)
 {
 	if(fdwReason == DLL_PROCESS_ATTACH)
 	{
-		gFS->RemoveFile("SAMP\\addon\\addon.log");
+		if(gFS->FileExist("SAMP\\addon\\addon.log"))
+			gFS->RemoveFile("SAMP\\addon\\addon.log");
 
 		addonDebug("\tDebugging started\n");
 		addonDebug("-----------------------------------------------------------------");
-		addonDebug("Called asi attach | hinstDLL 0x%x | fdwReason %i | lpvReserved %i", hinstDLL, fdwReason, lpvReserved);
+		addonDebug("Called asi attach | asi base address: 0x%x | attach reason: %i | reserved: %i", hinstDLL, fdwReason, lpvReserved);
 		addonDebug("-----------------------------------------------------------------");
+
+		if(!gFS->FileExist("addon.asi") || (GetModuleHandleW(L"addon.asi") != hinstDLL))
+		{
+			addonDebug("Addon loader must be named as 'addon.asi'. Please, rename it");
+
+			return false;
+		}
 
 		gFS = new addonLoaderFS();
 		std::vector<std::string> asidel = gFS->ListDirectory("./");
@@ -36,26 +44,31 @@ BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved)
 			if((*i).find(".asi") != ((*i).length() - 4))
 				continue;
 
-			addonDebug("Removing %s due to asi block activated", (*i).c_str());
+			addonDebug("Removing '%s' due to asi block activated", (*i).c_str());
 
 			gFS->RemoveFile((*i));
 		}
 
-		if(!GetModuleHandleW(L"gta_sa.exe") || !GetModuleHandleW(L"VorbisFile.dll") || !GetModuleHandleW(L"VorbisHooked.dll"))
+		if(!GetModuleHandleW(L"gta_sa.exe") || !GetModuleHandleW(L"Vorbis.dll") || !GetModuleHandleW(L"VorbisFile.dll") || !GetModuleHandleW(L"VorbisHooked.dll"))
 		{
 			addonDebug("Loader attached from unknown process, terminating");
 
 			return false;
 		}
 
-		boost::thread load(&addon_load_thread);
+		addonDebug("Game process base address: 0x%x", GetModuleHandleW(L"gta_sa.exe"));
+		addonDebug("Vorbis library base address: 0x%x", GetModuleHandleW(L"Vorbis.dll"));
+		addonDebug("VorbisFile hook base address: 0x%x", GetModuleHandleW(L"VorbisFile.dll"));
+		addonDebug("VorbisFile library base address: 0x%x", GetModuleHandleW(L"VorbisHooked.dll"));
+
+		boost::thread load(boost::bind(&addon_load_thread));
 	}
 	else if(fdwReason == DLL_PROCESS_DETACH)
 	{
 		delete gFS;
 
 		addonDebug("-----------------------------------------------------------------");
-		addonDebug("Called asi detach | hinstDLL 0x%x | fdwReason %i | lpvReserved %i", hinstDLL, fdwReason, lpvReserved);
+		addonDebug("Called asi detach | asi base address: 0x%x | detach reason: %i | reserved: %i", hinstDLL, fdwReason, lpvReserved);
 		addonDebug("-----------------------------------------------------------------\n");
 		addonDebug("\tDebugging stopped");
 	}
@@ -71,14 +84,14 @@ void addon_load_thread()
 
 	while(!GetModuleHandleW(L"samp.dll"))
 	{
-		if(++processTicks >= 120) // 30 seconds
+		if(++processTicks >= 30) // 30 seconds
 		{
 			addonDebug("Singleplayer loaded, terminating");
 
 			return;
 		}
 
-		boost::this_thread::sleep(boost::posix_time::milliseconds(250));
+		boost::this_thread::sleep(boost::posix_time::seconds(1));
 	}
 
 	SetDllDirectoryW(L"SAMP\\addon");
@@ -121,14 +134,14 @@ void addonDebug(char *text, ...)
 	va_start(args, text);
 
 	int length = vsnprintf(NULL, NULL, text, args);
-	char *buffer = (char *)malloc(++length);
+	char *buffer = new char[++length];
 
 	vsnprintf(buffer, length, text, args);
 	va_end(args);
 
 	logfile.open("SAMP\\addon\\addon.log", (std::fstream::out | std::fstream::app));
-	logfile << '[' << timeform << ']' << ' ' << buffer << std::endl;
+	logfile << "[" << timeform << "] " << buffer << std::endl;
 	logfile.close();
 
-	free(buffer);
+	delete[] buffer;
 }
